pyproject.toml
[project]
name = "567-Project"
version = "0.0.0"
requires-python = ">=3.14"
dependencies = [
        "jupyter", "numpy", "matplotlib",
        "tqdm",
        "wandb",
        "torch>=2.9.1",
        "torchrl>=0.10, <0.12",
        "torchvision",
        "pytorch-triton-rocm>=3.5.1 ; sys_platform == 'linux'",
        "benchmarl[vmas]>=1.5.2", "pettingzoo[classic,butterfly,mpe,sisl,other]",
        # "pyglet",
	"id-marl-eval",
]

[tool.uv]
  override-dependencies = ["pandas==3.0.1", "arch>=8.0"]

[tool.uv.sources]
torch = [
  { index = "pytorch-rocm", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-rocm", marker = "sys_platform == 'linux'" },
]
# ROCm6 support relies on `pytorch-triton-rocm`, which should also be installed from the PyTorch index
# (and included in `project.dependencies`).
pytorch-triton-rocm = [
  { index = "pytorch-rocm", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm6.4"
explicit = true
